+++ js/engine_struct.js
@@ -1070,10 +1070,67 @@ class AggregatedResult {
   }
 }
 
+/**
+ * Class representing the results of a simulation run.
+ */
+class SimulationResult {
+  /**
+   * Creates a new simulation result instance
+   * @param {string} name - The name of the simulation
+   * @param {Array} trialResults - Array containing the results of each trial
+   *     run.
+   */
+  constructor(name, trialResults) {
+    const self = this;
+    self._name = name;
+    self._trialResults = trialResults;
+  }
+
+  /**
+   * Gets the name of the simulation as defined in code.
+   *
+   * @returns {string} The simulation name.
+   */
+  getName() {
+    const self = this;
+    return self._name;
+  }
+
+  /**
+   * Gets the results from all trial runs.
+   *
+   * @returns {Array<EngineResult>} Array of trial results.
+   */
+  getTrialResults() {
+    const self = this;
+    return self._trialResults;
+  }
+}
+
+class SimulationAttributeToExporterResults {
+  constructor(inner) {
+    const self = this;
+    self._inner = inner;
+    self._trialResults = self._inner.getTrialResults().map((x) => new AttributeToExporterResult(x));
+  }
+
+  getName() {
+    const self = this;
+    return self._inner.getName();
+  }
+
+  getTrialResults() {
+    const self = this;
+    return self._trialResults;
+  }
+}
+
 export {
   AggregatedResult,
   AttributeToExporterResult,
   EngineResult,
   EngineResultBuilder,
   ImportSupplement,
+  SimulationResult,
+  SimulationAttributeToExporterResults,
 };
